refactor(web): use dynamic import for subcommand
