+++ packages/opencode/src/index.ts
@@ -26,7 +26,7 @@ import { AttachCommand } from "./cli/cmd/tui/attach/command"
 import { TuiThreadCommand } from "./cli/cmd/tui/thread/command"
 import { AcpCommand } from "./cli/cmd/acp/command"
 import { EOL } from "os"
-import { WebCommand } from "./cli/cmd/web"
+import { WebCommand } from "./cli/cmd/web/command"
 import { PrCommand } from "./cli/cmd/pr/command"
 import { SessionCommand } from "./cli/cmd/session/command"
 import { DbCommand } from "./cli/cmd/db/command"
